src/backend/requirements.txt
flask==2.3.2
pandas==2.0.2
requests==2.31.0
python-dotenv==1.0.0
openpyxl==3.1.2